requirements.txt
python-telegram-bot==20.7
python-decouple==3.8
httpx==0.28.1
beautifulsoup4==4.14.2
lxml==6.0.2

# Additional dependencies
python-dateutil==2.9.0
rapidfuzz==3.10.1  # Fast fuzzy string matching

# FastAPI Landing Server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
